optee: add FF-A support
Adds support for using FF-A [1] as transport to the OP-TEE driver.

Introduces struct optee_msg_param_smem which carries all information
needed when OP-TEE is calling FFA_MEM_RETRIEVE_REQ to get the shared
memory reference mapped by the hypervisor in S-EL2. Register usage is
also updated to include the information needed.

This isn't strictly needed without a hypervisor in S-EL2 and OP-TEE
itself has the SPMC part. However, this is a configuration which
shouldn't concern normal world.

Signed-off-by: Jens Wiklander <jens.wiklander@linaro.org>