arm_spci: Expose SPCI_PARTITION_INFO_GET
Expose the SPCI_PARTITION_INFO_GET ABI to allow querying the system for
information about a set of partitions identified by a specified UUID.

The function expects the 128bit UUID to be discovered passed via 4 32bit
words in the format: UUID_0-UUID_1-UUID_2-UUID3 and a pointer to an
array of `spci_partition_info` structs that will allocated and populated
with the discovered partitions information by the SPCI driver.  The
caller is responsible for freeing this memory allocated by the SPCI
driver.

The function returns the number of the partitions found (and therefore
the number of elements in the array) or a negative value indicating the
error code if unsuccessful.

This implementation currently disallows use of the NULL UUID.

Change-Id: I8239dd431199a4c30e861496005407b9c7a6e2f5
Signed-off-by: Jose Marinho <jose.marinho@arm.com>