arm_spci: Add support for non-Tx buffer mem_share
This commit introduces the buffer*, buffer_size parameters to the
mem_share function. The buffer can be used alternatively to the Tx
buffer. The mem_share function will populate the IPA-contiguous memory
pointed to by buffer using the information in sg.

The buffer* must be NULL is mem_share should use the Tx buffer.
The buffer*, if non-NULL, must point to a IPA-contiguous memory region
of size multiple of 4kiB.

Change-Id: I33f070eec899d849e97a2cec10039cc48757bd44