arm_spci: mem_share selection of Tx or alloc buf
This commit allows the mem_share invokation to select the  tx buffer or a
dynamically allocated buffer to be used for the transfer of the memory
region description.

if use_tx == true: the tx_lock is held until the memory region is fully
transmitted.

if use_tx == false: the tx_lock is very briefly held in the preamble
and postamble of the mem_share procedure.

Change-Id: I4e86334ff19f570e0d83d7b047e6fe3731a10474
Signed-off-by: Jose Marinho <jose.marinho@arm.com>