optee: updated for FF-A 1.0
With FF-A 1.0 the ABI to secure world has changed compared to SPCI
beta1.

The globally unique handle to identify a shared memory object is now 64
bits wide instead of previous 32. The IDR in struct optee_ffa is
replaced with a struct rhashtable instead in order to support 64-bit
keys.

Register usage in the FF-A function has changed compared to SPCI beta1
so the OP-TEE ABI is updated accordingly. The global handle, aka cookie,
is now requires 2 registers.

struct optee_msg_param_smem is updated to be able to carry a 64-bit
global handle while maintaining the same size of the struct. Due to this
the "page_count" field is dropped.

Signed-off-by: Jens Wiklander <jens.wiklander@linaro.org>