optee: sync optee_msg.h and optee_rpc_cmd.h
Updates to latest optee_msg.h and optee_rpc_cmd.h. There's no changes in
the ABI. Only some clarifications and a complete specification of RPC
requests where the latter is now in a separate file, optee_rpc_cmd.h.

Most of the RPC requests are not served by the OP-TEE driver instead
they are forwarded as opaque requests to tee-supplicant.

Signed-off-by: Jens Wiklander <jens.wiklander@linaro.org>