ffa_driver: Add Notification mechanism
Add interface to the FFA driver to allow for client
drivers to request and relinquish a notification and
provide a callback for the assigned notification.
Notification IDs are assigned by the lowest ID that is
not currently already assigned.

This patch also registers an internal handlers for
handling notifications targeted to the FF-A driver
directly.

Signed-off-by: Marc Bonnici <marc.bonnici@arm.com>