ffa_driver: Implement NOTIFICATION_BITMAP_CREATE
On systems without a hypervisor the responsibility
of requesting the creation of the notification bitmaps
in the SPM falls to the FF-A driver.

We use FFA Features to determine if the ABI is
supported, if it is not we can assume there is a
hypervisor present and will take care of ensure the
relevant notifications bitmaps are created on this
partitions behalf.

Signed-off-by: Marc Bonnici <marc.bonnici@arm.com>