firmware: arm_ffa: Use 64bit RXTX_MAP message
ffa_rxtx_map() was always using the 32bit RXTX_MAP FF-A message when
trying to map the rxtx buffer in the secure world. This resulted in a
possible incorrect pointer being passed to the SPM. The SPMD only
passes the lower 32bit address of the pointer when the 32bit message
is used.

Signed-off-by: Jelle Sels <jelle.sels@arm.com>
Signed-off-by: Sudeep Holla <sudeep.holla@arm.com>