test_ffa_driver: Add Test Driver
Add an test driver on top of the FF-A Driver to exercise
FFA notification support and demonstrate the expected interface
between a client driver and the FF-A Driver.

This driver requires a userspace application to be added
to the linux filesystem to allow communication with the driver.
It also requires a Hafnium bare metal partition loaded as multiple
VMs and SPs in the system to demonstrate inter-partition
communication.

Signed-off-by: Marc Bonnici <marc.bonnici@arm.com>