arm64: dt: rockpi4: disable node usbdrd_dwc3_0 **not for mainline**
If a USB-A to USB-A cable is used to flash the RockPi4 eMMC as described
in [1], it is likely that the board will be booted with the cable still
plugged into the board and into the computer on the other side. Such a
configuration results in periodic error messages from the kernel:

 [    4.832697] usb usb6-port1: Cannot enable. Maybe the USB cable is bad?
 [    4.833416] usb usb6-port1: config error

That is annoying, especially since the messages are also sent to the
console by default, which could mess up with CI scripts. Therefore,
disable this port (it is the upper USB3 blue port; the lower one will
still work fine).

Link: [1] https://wiki.radxa.com/Rockpi4/dev/usb-install
Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Acked-by: Jens Wiklander <jens.wiklander@linaro.org>
Acked-by: Joakim Bech <joakim.bech@linaro.org>